Replace generic prebuilt bindings with target specific bindings
* Replace generic prebuilt bindings with target specific bindings

This commit replaces the generic gdal bindings with architecture
specific ones. This is necessary as bindgen 0.71 introduced a feature
that essentially turned the size assertions into a compile time check.
This pointed to several potential mismatches between the bindings and
the actual type definitions.

This commit now introduces different bindings for 32 and 64 bit
platforms. Additionally it differentiates between windows and
linux/macos bindings. For any unsupported platform a build error is
generated that the built_time bindgen feature needs to be used there.

Finally it adds documentation on how to generate these bindings.